LPS-56703 Move Command classes to OSGi module and convert them to OSGi components
