+++ portal-impl/src/portal.properties
@@ -831,7 +831,6 @@
         META-INF/messaging-misc-spring.xml,\
         \
         META-INF/audit-spring.xml,\
-        META-INF/blogs-spring.xml,\
         META-INF/cluster-spring.xml,\
         META-INF/comment-spring.xml,\
         META-INF/document-library-spring.xml,\
